Update of plotStocks.py.  Added ability to select stocks based on user feedback.
