+++ ideas.txt
@@ -1,0 +1,1 @@
+/Users/jcanepa/Documents/Seniors Nigh.rtf
